refactor: streamline app bottom bar widget
Streamline app bottom bar to garden, asks, and settings
