+++ plural_app/lib/src/features/asks/presentation/listed_asks_dialog.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-import 'package:intl/intl.dart';
 
 // Common Widgets
-import 'package:plural_app/src/common_widgets/app_dialog.dart';
 import 'package:plural_app/src/common_widgets/app_dialog_footer.dart';
 
 // Constants
 import 'package:plural_app/src/constants/app_sizes.dart';
-import 'package:plural_app/src/constants/formats.dart';
 
 // Asks
-import 'package:plural_app/src/features/asks/data/asks_api.dart';
 import 'package:plural_app/src/features/asks/presentation/listed_ask_tile.dart';
 
 // Localization
@@ -19,30 +15,6 @@ import 'package:plural_app/src/localization/lang_en.dart';
 
 // Utils
 import 'package:plural_app/src/utils/app_dialog_router.dart';
-import 'package:plural_app/src/utils/app_state.dart';
-
-Future createListedAsksDialog(BuildContext context) async {
-  final currentUserID = GetIt.instance<AppState>().currentUserID!;
-  final nowString = DateFormat(Formats.dateYMMddHms).format(DateTime.now());
-
-  final asks = await getAsksForListedAsksDialog(
-    userID: currentUserID,
-    nowString: nowString
-  );
-
-  if (context.mounted) {
-    return showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AppDialog(
-          view: AskDialogList(
-            listedAskTiles: [for (var ask in asks) ListedAskTile(ask: ask)]
-          ),
-        );
-      }
-    );
-  }
-}
 
 class AskDialogList extends StatelessWidget {
   const AskDialogList({
@@ -68,12 +40,12 @@ class AskDialogList extends StatelessWidget {
         AppDialogFooterBuffer(buttons: [RouteToCreateAskViewButton()],),
         AppDialogNavFooter(
           leftDialogIcon: Icons.local_florist,
-          leftNavCallback: appDialogRouter.routeToGardenDialogListView,
-          leftTooltipMessage: AppDialogFooterText.navToGardens,
+          leftNavCallback: appDialogRouter.routeToCurrentGardenDialogView,
+          leftTooltipMessage: AppDialogFooterText.navToGardenDialog,
           rightDialogIcon: Icons.settings,
           rightNavCallback: appDialogRouter.routeToUserSettingsDialogView,
-          rightTooltipMessage: AppDialogFooterText.navToSettings,
-          title: AppDialogFooterText.asks
+          rightTooltipMessage: AppDialogFooterText.navToSettingsDialog,
+          title: AppDialogFooterText.listedAsks
         )
       ],
     );
