+++ plural_app/test/features/asks/presentation/creatable_ask_dialog_test.dart
@@ -5,7 +5,7 @@ import 'package:get_it/get_it.dart';
 // Common Widgets
 import 'package:plural_app/src/common_widgets/app_currency_picker_form_field.dart';
 import 'package:plural_app/src/common_widgets/app_date_picker_form_field.dart';
-import 'package:plural_app/src/common_widgets/app_dialog.dart';
+import 'package:plural_app/src/common_widgets/app_dialog_footer.dart';
 import 'package:plural_app/src/common_widgets/app_dialog_footer_buffer_submit_button.dart';
 import 'package:plural_app/src/common_widgets/app_text_form_field.dart';
 
@@ -35,16 +35,29 @@ void main() {
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: AppDialog(view: AskDialogCreateForm())
+            body: Builder(
+              builder: (BuildContext context) {
+                return ElevatedButton(
+                  onPressed: () => createCreatableAskDialog(context),
+                  child: Text("The ElevatedButton")
+                );
+              }
+            )
           ),
         )
       );
+
+      // Tap ElevatedButton (to open dialog)
+      await tester.tap(find.byType(ElevatedButton));
+      await tester.pumpAndSettle();
 
       // Check expected values are found
       expect(find.byType(AppDatePickerFormField), findsOneWidget);
       expect(find.byType(AppTextFormField), findsNWidgets(5));
       expect(find.byType(AppCurrencyPickerFormField), findsOneWidget);
       expect(find.byType(AppDialogFooterBufferSubmitButton), findsOneWidget);
+      expect(find.byType(AppDialogFooterBuffer), findsOneWidget);
+      expect(find.byType(AppDialogNavFooter), findsOneWidget);
     });
 
     tearDown(() => GetIt.instance.reset());
