+++ plural_app/test/features/asks/presentation/listed_asks_dialog_test.dart
@@ -1,20 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
-import 'package:mocktail/mocktail.dart';
-import 'package:pocketbase/pocketbase.dart';
 
 // Common Widgets
+import 'package:plural_app/src/common_widgets/app_dialog.dart';
 import 'package:plural_app/src/common_widgets/app_dialog_footer.dart';
 
 // Asks
-import 'package:plural_app/src/features/asks/data/asks_repository.dart';
 import 'package:plural_app/src/features/asks/presentation/creatable_ask_dialog.dart';
 import 'package:plural_app/src/features/asks/presentation/listed_ask_tile.dart';
 import 'package:plural_app/src/features/asks/presentation/listed_asks_dialog.dart';
-
-// Auth
-import 'package:plural_app/src/features/authentication/data/users_repository.dart';
 
 // Utils
 import 'package:plural_app/src/utils/app_dialog_router.dart';
@@ -22,7 +17,6 @@ import 'package:plural_app/src/utils/app_state.dart';
 
 // Tests
 import '../../../test_context.dart';
-import '../../../test_mocks.dart';
 
 void main() {
   group("AskDialogList test", () {
@@ -35,53 +29,22 @@ void main() {
 
       // GetIt
       final getIt = GetIt.instance;
-      final mockAsksRepository = MockAsksRepository();
-      final mockUsersRepository = MockUsersRepository();
       getIt.registerLazySingleton<AppState>(() => appState);
       getIt.registerLazySingleton<AppDialogRouter>(() => AppDialogRouter());
-      getIt.registerLazySingleton<AsksRepository>(() => mockAsksRepository);
-      getIt.registerLazySingleton<UsersRepository>(() => mockUsersRepository);
-
-      // AsksRepository.getAsksByUserID()
-      when(
-        () => mockAsksRepository.getList(
-          filter: any(named: "filter"),
-          sort: any(named: "sort"),
-        )
-      ).thenAnswer(
-        (_) async => ResultList<RecordModel>(items: [tc.getAskRecordModel()])
-      );
-
-      // UsersRepository.getFirstListItem()
-      when(
-        () => mockUsersRepository.getFirstListItem(
-          filter: any(named: "filter")
-        )
-      ).thenAnswer(
-        (_) async => tc.getUserRecordModel()
-      );
 
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: Builder(
-              builder: (BuildContext context) {
-                return ElevatedButton(
-                  onPressed: () => createListedAsksDialog(context),
-                  child: Text("The ElevatedButton")
-                );
-              }
+            body: AppDialog(
+              view: AskDialogList(listedAskTiles: [
+                ListedAskTile(ask: tc.ask),
+                ListedAskTile(ask: tc.ask),
+                ListedAskTile(ask: tc.ask),
+              ])
             )
           ),
         )
       );
-
-      // Check AskDialogList not yet displayed
-      expect(find.byType(AskDialogList), findsNothing);
-
-      // Tap ElevatedButton (to open dialog)
-      await tester.tap(find.byType(ElevatedButton));
-      await tester.pumpAndSettle();
 
       // Check expected values are found
       expect(find.byType(AskDialogList), findsOneWidget);
@@ -113,43 +76,18 @@ void main() {
 
       // GetIt
       final getIt = GetIt.instance;
-      final mockAsksRepository = MockAsksRepository();
       getIt.registerLazySingleton<AppState>(() => appState);
       getIt.registerLazySingleton<AppDialogRouter>(() => AppDialogRouter());
-      getIt.registerLazySingleton<AsksRepository>(() => mockAsksRepository);
-
-      // AsksRepository.getList()
-      when(
-        () => mockAsksRepository.getList(
-          filter: any(named: "filter"),
-          sort: any(named: "sort"),
-        )
-      ).thenAnswer(
-        (_) async => ResultList<RecordModel>(items: [])
-      );
 
       await tester.pumpWidget(
         MaterialApp(
           home: Scaffold(
-            body: Builder(
-              builder: (BuildContext context) {
-                return ElevatedButton(
-                  onPressed: () => createListedAsksDialog(context),
-                  child: Text("The ElevatedButton")
-                );
-              }
+            body: AppDialog(
+              view: AskDialogList(listedAskTiles: [])
             )
           ),
         )
       );
-
-      // Check widgets not yet displayed
-      expect(find.byType(AskDialogList), findsNothing);
-      expect(find.byType(EmptyListedAskTilesMessage), findsNothing);
-
-      // Tap ElevatedButton (to open dialog)
-      await tester.tap(find.byType(ElevatedButton));
-      await tester.pumpAndSettle();
 
       // Check expected values are found
       expect(find.byType(AskDialogList), findsOneWidget);
